+++ omibio/utils/check_if_exist.py
@@ -1,0 +1,4 @@
+def check_if_exist(string: str | None, default: str = "N/A") -> str:
+    if not string:
+        return default
+    return string