+++ omibio/utils/check_if_exsit.py
@@ -1,4 +1,0 @@
-def check_if_exsit(string: str | None, default: str = "N/A") -> str:
-    if not string:
-        return default
-    return string